CI build errors

+++ libext/imgui-SFML.cpp
@@ -374,7 +374,7 @@ namespace
                 if (pcmd->UserCallback) {
                     pcmd->UserCallback(cmd_list, pcmd);
                 } else {
-                    glBindTexture(GL_TEXTURE_2D, (GLuint)(unsigned int)pcmd->TextureId);
+                    glBindTexture(GL_TEXTURE_2D, (GLuint)(unsigned int)(size_t)pcmd->TextureId);
                     glScissor((int)pcmd->ClipRect.x, (int)(fb_height - pcmd->ClipRect.w),
                               (int)(pcmd->ClipRect.z - pcmd->ClipRect.x), (int)(pcmd->ClipRect.w - pcmd->ClipRect.y));
                     glDrawElements(GL_TRIANGLES, (GLsizei)pcmd->ElemCount, GL_UNSIGNED_SHORT, idx_buffer);
